+++ temba/users/models.py
@@ -2,7 +2,6 @@ import pyotp
 
 from django.conf import settings
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin, UserManager as AuthUserManager
-from django.contrib.auth.validators import UnicodeUsernameValidator
 from django.core.files.storage import storages
 from django.db import models
 from django.utils import timezone
@@ -51,12 +50,30 @@ class UserManager(AuthUserManager):
     Overrides the default user manager to make username lookups case insensitive
     """
 
-    def get_by_natural_key(self, username):
-        return self.get(**{f"{self.model.USERNAME_FIELD}__iexact": username})
+    def get_by_natural_key(self, email: str):
+        return self.get(**{f"{self.model.USERNAME_FIELD}__iexact": email})
+
+    def create_user(self, email: str, password: str, **extra_fields):
+        """
+        Create and save a user with the given email and password.
+        """
+        email = self.normalize_email(email)
+        user = self.model(email=email, **extra_fields)
+        user.set_password(password)
+        user.save()
+        return user
+
+    def create_system_user(self):
+        """
+        Creates the system user
+        """
+        user = self.model(email=User.SYSTEM["email"], first_name=User.SYSTEM["first_name"], is_system=True)
+        user.save()
+        return user
 
 
 class User(AbstractBaseUser, PermissionsMixin):
-    SYSTEM = {"username": "system", "first_name": "System"}
+    SYSTEM = {"email": "system", "first_name": "System"}
 
     STATUS_UNVERIFIED = "U"
     STATUS_VERIFIED = "V"
@@ -68,22 +85,15 @@ class User(AbstractBaseUser, PermissionsMixin):
     )
 
     EMAIL_FIELD = "email"
-    USERNAME_FIELD = "username"
-    REQUIRED_FIELDS = ["email"]
+    USERNAME_FIELD = "email"
+    REQUIRED_FIELDS = []
 
-    username = models.CharField(
-        _("username"),
-        max_length=150,
-        unique=True,
-        help_text=_("Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only."),
-        validators=[UnicodeUsernameValidator()],
-        error_messages={
-            "unique": _("A user with that username already exists."),
-        },
-    )
+    # username is on its way out
+    username = models.CharField(_("username"), max_length=150, null=True)
+
     first_name = models.CharField(_("first name"), max_length=150, blank=True)
     last_name = models.CharField(_("last name"), max_length=150, blank=True)
-    email = models.EmailField(_("email address"), blank=True)
+    email = models.EmailField(_("email address"), unique=True)
     language = models.CharField(max_length=8, choices=settings.LANGUAGES, default=settings.DEFAULT_LANGUAGE)
     avatar = models.ImageField(upload_to=UploadToIdPathAndRename("avatars/"), storage=storages["public"], null=True)
 
@@ -145,7 +155,7 @@ class User(AbstractBaseUser, PermissionsMixin):
 
     @classmethod
     def get_by_email(cls, email: str):
-        return cls.objects.filter(username__iexact=email).first()
+        return cls.objects.filter(email__iexact=email).first()
 
     @classmethod
     def get_orgs_for_request(cls, request):
@@ -160,14 +170,7 @@ class User(AbstractBaseUser, PermissionsMixin):
         """
         Gets the system user
         """
-        return cls.objects.get(username=cls.SYSTEM["username"])
-
-    @classmethod
-    def _create_system_user(cls):
-        """
-        Creates the system user
-        """
-        cls.objects.create_user(cls.SYSTEM["username"], first_name=cls.SYSTEM["first_name"], is_system=True)
+        return cls.objects.get(email=cls.SYSTEM["email"])
 
     @property
     def name(self) -> str:
@@ -272,11 +275,9 @@ class User(AbstractBaseUser, PermissionsMixin):
         """
         Releases this user, and any orgs of which they are the sole owner.
         """
-        user_uuid = str(uuid4())
         self.first_name = ""
         self.last_name = ""
-        self.email = f"{user_uuid}@rapidpro.io"
-        self.username = f"{user_uuid}@rapidpro.io"
+        self.email = str(uuid4())
         self.password = ""
         self.is_active = False
         self.save()
@@ -293,7 +294,7 @@ class User(AbstractBaseUser, PermissionsMixin):
             org.remove_user(self)
 
     def __str__(self):
-        return self.name or self.username
+        return self.name or self.email
 
     class Meta:
         verbose_name = _("user")